refactor: sns-page sns-page에서 기존의 ui를 제거하고 데이터 통신과 기능 작동 방식에 맞게 새로운 TrackingImageActions ui를 작성했습니다.
