+++ src/1_Page/Sns/ui/TrackingImageActions/style.js
@@ -1,0 +1,22 @@
+import styled from "styled-components";
+
+const STYLE = {
+  InfoContainer: styled.div`
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 0 4px;
+    font-size: medium;
+  `,
+  IconContainer: styled.div`
+    display: flex;
+    justify-content: center;
+    gap: 4px;
+  `,
+  Icon: styled.img`
+    width: 24px;
+    cursor: pointer;
+  `
+};
+
+export default STYLE;
